Create app with multiple outs and a static class
This app takes user input to divide two numbers. If the user doesn't put a divisor in, it defaults to 2. The app also has a static class that just informs the user what a prime number is, and then states that it isn't prime, or that it might be prime.